+++ cucumber/authenticators/features/step_definitions/authn_ldap_steps.rb
@@ -1,32 +1,24 @@
 # frozen_string_literal: true
 
-# Uses cucumber's multiline string feature: https://bit.ly/2vpzqJx
-#
-
-When(/I login via LDAP as (?:\S)+ Conjur user "(\S+)"/) do |username|
-  login_with_ldap(service_id: 'test', account: 'cucumber', 
-                  username: username, password: username)
-end
-
 # First non-captured group allows for adjectives to clarify the
 # purpose of the test.  They aren't actually used.
 #
-When(/I authenticate via LDAP as (?:\S)+ Conjur user "(\S+)"( using key)?/) do |username, using_key|
-  password = using_key ? ldap_auth_key : username
+When(/I authenticate via LDAP as (?:\S)+ Conjur user "(\S+)"/) do |username|
+  password = username
   authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
-                         username: username, api_key: password)
+                         username: username, password: password)
 end
 
 When(/my LDAP password for (?:\S)+ Conjur user "(\S+)" is empty/) do |username|
-  login_with_ldap(service_id: 'test', account: 'cucumber', 
-                  username: username, password: '')
+  password = ""
+  authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
+    username: username, password: password)
 end
 
-
-
 When(/my LDAP password is wrong for authorized user "(\S+)"/) do |username|
-  login_with_ldap(service_id: 'test', account: 'cucumber', 
-                  username: username, password: 'BAD_PASSWORD')
+  password = "bad_password"
+  authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
+    username: username, password: password)
 end
 
 Then(/it is denied/) do